upload

+++ app/Http/Controllers/admin/categoryController.php
@@ -30,7 +30,7 @@ class categoryController extends Controller
                 ->orWhere('slug', 'like', '%' . $request->search . '%');
         }
         $currentPage = $request->input('page', 1);
-
+        Session::put('page', $currentPage);
         $categories = $categories->paginate(6, ['*'], 'page', $currentPage);
         return view('admin.category.list_data', compact('categories'))->render();
     }
@@ -66,8 +66,7 @@ class categoryController extends Controller
             unset($input['_token']);
             unset($input['page']);
             Category::create($input);
-            $url = url('admin/category/index/' .$request->page );
-            dd($url);
+            $url = url('admin/category/index') . '?page=' . Session::get('page');
             return response()->json([
                 'status' => 1,
                 'message' => 'Thêm danh mục thành công',
@@ -81,33 +80,76 @@ class categoryController extends Controller
         }
     }
 
-    public function edit($slug): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
+    public function edit($id): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
     {
-        $categories = Category::where('slug', $slug)->first();
-        $category = Category::all();
-        return view('admin.category.edit', compact('categories','category'));
+        $category = Category::find($id);
+        $categories = Category::all();
+        return view('admin.category.modal.edit', compact('categories','category'));
     }
 
 
-    public function update(Request $request, $id, CategoryRequest $categoryRequest): \Illuminate\Http\RedirectResponse
+    public function update(Request $request)
     {
-        $categoryRequest->validated();
+        $id = $request->id;
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|max:255',
+            'slug' => 'required|max:255',
+        ], [
+            'name.required' => 'Tên danh mục không được để trống',
+            'name.max' => 'Tên danh mục không được quá 255 ký tự',
+            'slug.required' => 'Slug không được để trống',
+            'slug.max' => 'Slug không được quá 255 ký tự',
+        ]);
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => 0,
+                'message' => $validator->errors()->toArray(),
+            ]);
+        }
         try {
-            $categories = Category::find($id);
+            $category = Category::find($id);
             $input = $request->all();
             unset($input['_token']);
-            $categories->update($input);
-            return redirect()->route('admin.category.index')->with('success', 'Đã sửa danh mục thành công');
+            unset($input['page']);
+            $category->update($input);
+            $url = url('admin/category/index') . '?page=' . Session::get('page');
+            return response()->json([
+                'status' => 1,
+                'message' => 'Thêm danh mục thành công',
+                'url' => $url,
+            ]);
         } catch (Exception $e) {
-            return redirect()->back()->with('error', 'Đã xảy ra lỗi');
+            return response()->json([
+                'status' => 2,
+                'message' => 'Đã xảy ra lỗi',
+            ]);
         }
     }
 
 
-    public function destroy($id)
+    public function destroy(Request $request): \Illuminate\Http\JsonResponse
     {
+        $id = $request->id;
         $category = Category::find($id);
-        $category->delete();
-        return redirect()->route('admin.category.index')->with('success', 'Đã xóa danh mục thành công');
+        if (empty($category)) {
+           return response()->json([
+                'status' => 0,
+                'message' => 'Không tìm thấy danh mục',
+           ]);
+        }
+        try {
+            $category->delete();
+            $url = url('admin/category/index') . '?page=' . Session::get('page');
+            return response()->json([
+                'status' => 1,
+                'message' => 'Xóa danh mục thành công',
+                'url' => $url,
+            ]);
+        } catch (Exception $e) {
+            return response()->json([
+                'status' => 2,
+                'message' => 'Đã xảy ra lỗi',
+            ]);
+        }
     }
 }
